// dev3-frontend-dev/services/api/dev3Schemas.ts
export type User = {
  /**
   * @format date-time
   */
  updatedAt: string;
  /**
   * @format date-time
   */
  createdAt: string;
  isCensored: boolean;
  isActive: boolean;
  uid: string;
  accountType: string;
  roles: string[];
  username: string;
  nearWalletAccountId: string;
};

export type File = {
  /**
   * @format date-time
   */
  updatedAt: string;
  /**
   * @format date-time
   */
  createdAt: string;
  name: string;
  mime_type: string;
  url: string;
  key: string;
  owner: User;
};

export type Project = {
  /**
   * @format date-time
   */
  updatedAt: string;
  /**
   * @format date-time
   */
  createdAt: string;
  name: string;
  slug: string;
  logo: File;
  owner: User;
};